isolate sections of overlapping intervals

Given a set of intervals in a table, isolate sections of intervals that are overlapping with other in intervals (optionally, within groups). Returns a data.table that contains intervals which are mutually non-overlapping or exactly overlapping with other intervals (ie there are no partially overlapping intervals) (optionally within groups). Note that this doesn't just return the intersects; the original interval data is conserved such that for each interval/row in x, the return table has one or more non-overlapping intervals that together form the union of that original interval.

isolateoverlaps(
  x,
  interval_vars,
  group_vars = NULL,
  interval_vars_out = c("start", "end")
)

Arguments

x

A data.table containing a set of intervals.

interval_vars

A length-2 character vector denoting column names in x. these columns must be of the same class and be integer or IDate. The column name specifying the lower-bound column must be specified first.

group_vars

NULL, or a character vector denoting column names in x. These columns serve as grouping variables such that testing for overlaps and subsequent isolation only occur within categories defined by the combination of the group variables.

interval_vars_out

The desired column names of the interval columns in the return data.table. By default these columns will be generated to be named c("start","end"). If x contains columns with the same name as the desired output column names specified in interval_vars_out, the function will return in error to avoid naming confusion in the return table.

Value

A data.table with columns interval_vars_out which denote the start and stop period for each new interval. This return table also contains columns in x (including the original interval columns).

Details

All intervals are treated as closed (ie inclusive of the start and end values in the columns specified by interval_vars)

x is not copied but rather passed by reference to function internals but the order of this data.tables is restored on function completion or error.

See also

Examples

set.seed(23)
x2 <- data.table(addr_id=rep(1:4,each=3),
                exposure_start=rep(c(1L,7L,14L),times=4),
exposure_end=rep(c(7L,14L,21L),times=4),
exposure_value=c(rnorm(12))
)
x2z <- isolateoverlaps(x2,interval_vars=c("exposure_start","exposure_end"),group_vars=c("addr_id"))
x2z
#>     addr_id start end exposure_start exposure_end exposure_value
#>  1:       1     1   6              1            7     0.19321233
#>  2:       1     7   7              1            7     0.19321233
#>  3:       1     7   7              7           14    -0.43468211
#>  4:       1     8  13              7           14    -0.43468211
#>  5:       1    14  14              7           14    -0.43468211
#>  6:       1    14  14             14           21     0.91326710
#>  7:       1    15  21             14           21     0.91326710
#>  8:       2     1   6              1            7     1.79338809
#>  9:       2     7   7              1            7     1.79338809
#> 10:       2     7   7              7           14     0.99660511
#> 11:       2     8  13              7           14     0.99660511
#> 12:       2    14  14              7           14     0.99660511
#> 13:       2    14  14             14           21     1.10749049
#> 14:       2    15  21             14           21     1.10749049
#> 15:       3     1   6              1            7    -0.27808628
#> 16:       3     7   7              1            7    -0.27808628
#> 17:       3     7   7              7           14     1.01920549
#> 18:       3     8  13              7           14     1.01920549
#> 19:       3    14  14              7           14     1.01920549
#> 20:       3    14  14             14           21     0.04543718
#> 21:       3    15  21             14           21     0.04543718
#> 22:       4     1   6              1            7     1.57577959
#> 23:       4     7   7              1            7     1.57577959
#> 24:       4     7   7              7           14     0.21828845
#> 25:       4     8  13              7           14     0.21828845
#> 26:       4    14  14              7           14     0.21828845
#> 27:       4    14  14             14           21    -1.04653534
#> 28:       4    15  21             14           21    -1.04653534
#>     addr_id start end exposure_start exposure_end exposure_value
#x2b represents x2 when where exposure values in overlapping intervals have been averaged
x2b <- x2z[, list(exposure_value=mean(exposure_value)),by=c("addr_id","start","end")]